[MIG] product_brand :Updated module as per v9
