+++ product_brand/__openerp__.py
@@ -1,45 +1,25 @@
-# -*- encoding: utf-8 -*-
-###############################################################################
-# #                                                                           #
-# product_brand for Odoo #                                                    #
-# Copyright (C) 2009 NetAndCo (<http://www.netandco.net>). #                  #
-# Copyright (C) 2011 Akretion Benoît Guillot <benoit.guillot@akretion.com> #  #
-# Copyright (C) 2014 prisnet.ch Seraphine Lantible <s.lantible@gmail.com> #   #
-# Contributors                                                                #
-# Mathieu Lemercier, mathieu@netandco.net, #                                  #
-# Franck Bret, franck@netandco.net #                                          #
-# Seraphine Lantible, s.lantible@gmail.com, http://www.prisnet.ch             #
-# #                                                                           #
-# This program is free software: you can redistribute it and/or modify #      #
-# it under the terms of the GNU Affero General Public License as #            #
-# published by the Free Software Foundation, either version 3 of the #        #
-# License, or (at your option) any later version. #                           #
-# #                                                                           #
-# This program is distributed in the hope that it will be useful, #           #
-# but WITHOUT ANY WARRANTY; without even the implied warranty of #            #
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the #              #
-# GNU Affero General Public License for more details. #                       #
-# #                                                                           #
-# You should have received a copy of the GNU Affero General Public License #  #
-# along with this program. If not, see <http://www.gnu.org/licenses/>. #      #
-# #                                                                           #
-###############################################################################
-###############################################################################
-# Product Brand is an Openobject module wich enable Brand management for      #
-# products                                                                    #
-###############################################################################
+# -*- coding: utf-8 -*-
+# © 2009 NetAndCo (<http://www.netandco.net>).
+# © 2011 Akretion Benoît Guillot <benoit.guillot@akretion.com>
+# © 2014 prisnet.ch Seraphine Lantible <s.lantible@gmail.com>
+# © 2016-Today Serpent Consulting Services Pvt. Ltd.
+#    (<http://www.serpentcs.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
+
 {
     'name': 'Product Brand Manager',
-    'version': '8.0.0.1.0',
+    'version': '9.0.1.0.0',
     'category': 'Product',
-    'summary': 'Add brand to products',
+    'summary': "Product Brand Manager",
     'author': 'NetAndCo, Akretion, Prisnet Telecommunications SA'
-              ', MONK Software, Odoo Community Association (OCA)',
+              ', MONK Software, Odoo Community Association (OCA)'
+              ', SerpentCS Pvt. Ltd.',
     'license': 'AGPL-3',
     'depends': ['product'],
     'data': [
-        'product_brand_view.xml',
+        'views/product_brand_view.xml',
         'security/ir.model.access.csv'
     ],
-    'installable': True,
+    'installable': False,
+    'auto_install': False
 }
