+++ product_brand/models/__init__.py
@@ -6,4 +6,4 @@
 #    (<http://www.serpentcs.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-from . import models
+from . import product_brand
