+++ product_brand/views/product_brand_view.xml
@@ -1,18 +1,7 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  product_brand for Odoo
-  Copyright (C) 2009 NetAndCo <http://www.netandco.net>.
-  Copyright (C) 2011 Akretion Benoît Guillot <benoit.guillot@akretion.com>
-  Copyright (C) 2014 prisnet.ch Seraphine Lantible <s.lantible@gmail.com>
-  Copyright (C) 2015 Leonardo Donelli <http://www.wearemonk.com>
-  Contributors:
-  Mathieu Lemercier <mathieu@netandco.net>
-  Franck Bret <franck@netandco.net>
-  Seraphine Lantible <s.lantible@gmail.com>
-  Leonardo Donelli <donelli@webmonks.it>
--->
+<?xml version="1.0" encoding="UTF-8"?>
 <openerp>
     <data>
+
         <record id="product_brand_search_form_view" model="ir.ui.view">
             <field name="name">product.brand.search.form</field>
             <field name="model">product.brand</field>
@@ -129,21 +118,6 @@
             </field>
         </record>
 
-        <record id="product_search_form_view" model="ir.ui.view">
-            <field name="name">product.search.form</field>
-            <field name="model">product.product</field>
-            <field name="inherit_id" ref="product.product_search_form_view"/>
-            <field name="arch" type="xml">
-                <field name="categ_id" position="after">
-                    <field name="product_brand_id"/>
-                </field>
-                <group string='Group by...' position="inside">
-                    <filter string="Brand" name="groupby_brand" domain="[]"
-                            context="{'group_by' : 'product_brand_id'}"/>
-                </group>
-            </field>
-        </record>
-
         <record id="view_product_template_search_brand" model="ir.ui.view">
             <field name="name">product.template.search.brand</field>
             <field name="model">product.template</field>
@@ -151,11 +125,10 @@
             <field name="arch" type="xml">
                 <field name="name" position="after">
                     <field name="product_brand_id"/>
-                </field>
-                <group string='Group by...' position="inside">
                     <filter string="Brand" name="groupby_brand" domain="[]"
                             context="{'group_by' : 'product_brand_id'}"/>
-                </group>
+                    <separator/>
+                </field>
             </field>
         </record>
 
@@ -175,7 +148,7 @@
             <field name="model">product.template</field>
             <field name="inherit_id" ref="product.product_template_kanban_view" />
             <field name="arch" type="xml">
-                <xpath expr="//h4" position="after">
+                <xpath expr="//div[@class='oe_kanban_details']/strong[1]" position="after">
                     <div>
                         <a t-if="record.product_brand_id" type="action"
                            name="%(action_open_single_product_brand)d">
@@ -191,7 +164,7 @@
             <field name="model">product.product</field>
             <field name="inherit_id" ref="product.product_kanban_view" />
             <field name="arch" type="xml">
-                <xpath expr="//h4" position="after">
+                <xpath expr="//div[@class='o_kanban_tags_section']" position="after">
                     <div>
                         <a t-if="record.product_brand_id" type="open">
                             <field name="product_brand_id"/>
@@ -235,5 +208,6 @@
             id="menu_product_brand"
             action="action_product_brand"
             parent="base.menu_product"/>
+
     </data>
 </openerp>
